feat: move RvcManager to another file
